+++ Sources/Core/Presenter/Channels/ChannelsPresenter+Rx.swift
@@ -44,6 +44,18 @@ public extension Reactive where Base == ChannelsPresenter {
             .asDriver(onErrorJustReturn: .empty)
     }
     
+    /// Mutes a channel from a given `ChannelPresenter`.
+    /// - Parameter channelPresenter: a channel presenter.
+    func mute(_ channelPresenter: ChannelPresenter) -> Observable<MutedChannelResponse> {
+        channelPresenter.channel.rx.mute()
+    }
+    
+    /// Unmutes a channel from a given `ChannelPresenter`.
+    /// - Parameter channelPresenter: a channel presenter.
+    func unmute(_ channelPresenter: ChannelPresenter) -> Observable<EmptyData> {
+        channelPresenter.channel.rx.unmute()
+    }
+    
     /// Delete a channel and remove a channel presenter from items.
     ///
     /// - Parameters: