+++ Sources/UI/Channels/ChannelsViewController.swift
@@ -369,6 +369,9 @@ extension ChannelsViewController: UITableViewDataSource, UITableViewDelegate {
             return
         }
         
-        channelPresenter.channel.rx.stopWatching().flatMap { _ in self.presenter.rx.delete(channelPresenter) }.subscribe().disposed(by: disposeBag)
+        channelPresenter.channel.rx.stopWatching()
+            .flatMap { _ in self.presenter.rx.delete(channelPresenter) }
+            .subscribe()
+            .disposed(by: disposeBag)
     }
 }